moves serializer file into serialization module
