+++ src/pydase/server/web_server/sio_setup.py
@@ -10,7 +10,7 @@ from pydase.data_service.state_manager import StateManager
 from pydase.utils.helpers import get_object_attr_from_path_list
 from pydase.utils.logging import SocketIOHandler
 from pydase.utils.serialization.deserializer import Deserializer, loads
-from pydase.utils.serializer import SerializedObject, dump
+from pydase.utils.serialization.serializer import SerializedObject, dump
 
 logger = logging.getLogger(__name__)
