+++ src/pydase/utils/serialization/serializer.py
